Revise title-generation function
This commit revises the title-generation function to accept a Vec of
Strings.  This allows it to accept other arguments that are passed to
.TH (such as the date and version).

It also removes the capitalization, in part because not all of these
arguments are capitalized and in part because that degree of semantic
formatting (titles need to be capitalized) seems more appropriate for a
higher level crate like `man`.